requirements.txt
# --- Base de datos ---
mysql-connector-python>=8.0
python-dotenv>=1.0

# --- Procesamiento de datos ---
pandas>=2.0
numpy>=1.24

# --- Computer Vision ---
ultralytics>=8.0
opencv-python-headless>=4.8
scikit-learn>=1.3

# --- Audio ---
openai-whisper>=20230918

# --- Scraping ---
instaloader>=4.10

# --- Video download ---
yt-dlp>=2024.1

# --- API Backend ---
fastapi>=0.110
uvicorn>=0.27
bcrypt>=4.1
python-jose[cryptography]>=3.3
python-multipart>=0.0.6
pyyaml>=6.0
gunicorn>=21.2


# --- Labeling (opcional, si usas Label Studio local) ---
# label-studio>=1.23
